Added Skeleton Loader to Player Profile Page
